用户查看评论时，仍获取state=0与1的所有评论，由前端处理state=0的评论

+++ src/org/sousai/dao/MatchDao.java
@@ -80,7 +80,7 @@ public interface MatchDao {
 	 * 
 	 * @return
 	 */
-	int deleteMatches(Long[] matchIds);
+	int deleteMatches(Integer[] matchIds);
 	
 	/**
 	 * 模糊匹配某个字段
